test/eventdev: fix ethdev port id to 16-bit
port id in test case is update to unsigned int 16 bit, to
prevent compilation failure when RTE_MAX_ETHPORTS port is
set more than 255.

Fixes: fc8030eb8fe9 ("test/eventdev: add tests for eth Rx adapter APIs")
Cc: stable@dpdk.org

Acked-by: Nikhil Rao <nikhil.rao@intel.com>

+++ test/test/test_event_eth_rx_adapter.c
@@ -107,7 +107,7 @@ port_init(uint8_t port, struct rte_mempool *mp)
 static int
 init_ports(int num_ports)
 {
-	uint8_t portid;
+	uint16_t portid;
 	int retval;
 
 	default_params.mp = rte_pktmbuf_pool_create("packet_pool",
